Update and rename jd_festival.js to jd_nzmh.js
